Add HitAndMissProbability module to refactor the main orchestrator method
